move scrol area to dedsign system and style it like the native one
