+++ packages/design-system/src/index.ts
@@ -32,6 +32,7 @@ export * from "./components/radio";
 export * from "./components/checkbox";
 export * from "./components/component-card";
 export * from "./components/primitives/arrow-focus";
+export * from "./components/scroll-area";
 
 // Not aligned
